feat
RegisterActivity 추가
